+++ org.eclipse.scout.rt.shared/src/main/java/org/eclipse/scout/rt/shared/data/basic/NamedBitMaskHelper.java
@@ -236,7 +236,7 @@ public final class NamedBitMaskHelper {
     byte result = bitMaskFor(m_nextIndex);
     m_nextIndex++;
 
-    m_masksByBitName.put(bitName, Byte.valueOf(result));
+    m_masksByBitName.put(bitName, result);
     return result;
   }
